refactor names and add a test for the denoising
