+++ src/main/java/streams/hexmap/CameraMapping.java
@@ -1,0 +1,7 @@
+package streams.hexmap;
+
+/**
+ * Created by kbruegge on 2/13/17.
+ */
+public class CameraMapping {
+}
